+++ src/sem/usefulness.rs
@@ -2,10 +2,9 @@
 //! - https://github.com/rust-lang/rust/blob/d331cb710f0/compiler/rustc_mir_build/src/thir/pattern/usefulness.rs
 //! - https://github.com/rust-lang/rust/blob/d331cb710f0/compiler/rustc_mir_build/src/thir/pattern/deconstruct_pat.rs
 use super::error::SemanticError;
-use crate::{
-    syntax::{CaseArm, Pattern, PatternKind, RangeEnd},
-    ty::Type,
-};
+use crate::syntax::token::RangeEnd;
+use crate::syntax::tree::{CaseArm, Pattern, PatternKind};
+use crate::ty::Type;
 use std::{
     cell::Cell,
     cmp::{max, min},
